+++ sphinx/source/changelog.rst
@@ -4,6 +4,10 @@ Changelog
 1.0.2b1
 -------
 * Added a second source for the Skool Daze TZX file to `images.ini`
+* Removed the documentation sources from the Pyskool distribution (they can be
+  obtained from GitHub_)
+
+.. _GitHub: https://github.com/skoolkid/pyskool
 
 1.0.1 (2012-12-07)
 ------------------